backend/requirements.txt
langchain
langchain-ollama
langchain-community
langchain_openai
chromadb
fastapi
uvicorn
python-dotenv